% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{as.ccTable.new}
\alias{as.ccTable.new}
\title{Converts data into ca-co format (updated)}
\usage{
as.ccTable.new(cases, controls, ...)
}
\arguments{
\item{Cases}{number of cases}

\item{Controls}{number of controls}
}
\value{
NA
}
\description{
Creates new matrix represented as a table for a case control study
}
