% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{as.ccTable}
\alias{as.ccTable}
\title{Converts data into ca-co format}
\usage{
as.ccTable(eCases, eControls, neCases, neControls, ...)
}
\arguments{
\item{eCases}{number of exposed cases}

\item{eControls}{number of exposes controls}

\item{neCases}{number of non-exposed cases}

\item{neControls}{number of non-exposed controls}
}
\value{
NA
}
\description{
Creates new matrix represented as a table for a case control study
}
