% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{as.data.frame.ccTable}
\alias{as.data.frame.ccTable}
\title{Convert wide data format into long data format}
\usage{
\method{as.data.frame}{ccTable}(x, ...)
}
\arguments{
\item{x}{wide data format data frame}
}
\value{
long data format data frame
}
\description{
Takes in wide data format for closed cohort and case control studies and converts it into long format
}
