% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{as.mccTable}
\alias{as.mccTable}
\title{Converts data into matched ca-co format}
\usage{
as.mccTable(matchRatio, ...)
}
\arguments{
\item{matchRatio}{ratio used to calculate number of controls chosen for each case}
}
\value{
NA
}
\description{
Creates new matrix represented as a table for a case control study
}
