% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{as.rateTable.new}
\alias{as.rateTable.new}
\title{Converts data to rate table (updated)}
\usage{
as.rateTable.new(cases, person.time, ...)
}
\arguments{
\item{cases}{number of cases}

\item{person.time}{amount of person time}
}
\value{
NA
}
\description{
Creates new matrix represented as a table with person time
}
