% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{as.rateTable}
\alias{as.rateTable}
\title{Converts data to rate table}
\usage{
as.rateTable(eCases, ePT, neCases, nePT, ...)
}
\arguments{
\item{eCases}{number of exposed cases}

\item{ePT}{amount of exposed person-time}

\item{neCases}{number of not exposed Cases}

\item{nePT}{amount of unexposed person-time}
}
\value{
NA
}
\description{
Creates new matrix represented as a table with person time
}
