% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{as.riskTable.new}
\alias{as.riskTable.new}
\title{Converts data to new risk table}
\usage{
as.riskTable.new(cases, noncases, ...)
}
\arguments{
\item{Cases}{number of cases}

\item{NonCases}{number of non-cases}
}
\value{
NA
}
\description{
Creates new matrix represented as a table for a closed cohort (or cross-sectional)
}
