% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{as.riskTable}
\alias{as.riskTable}
\title{Converts data to risk table}
\usage{
as.riskTable(eCases, eNonCases, neCases, neNonCases, ...)
}
\arguments{
\item{eCases}{number of exposed cases}

\item{eNonCases}{number of exposes non-cases}

\item{neCases}{number of non-exposed cases}

\item{neNonCases}{number of non-exposed non-cases}
}
\value{
NA
}
\description{
Creates new matrix represented as a table for a closed cohort (or cross-sectional)
}
