% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{print.ccTable}
\alias{print.ccTable}
\title{Prints data from ca-co format}
\usage{
\method{print}{ccTable}(x, ...)
}
\arguments{
\item{x}{ccTable object}
}
\value{
NA
}
\description{
Creates new matrix represented as a table for a case control study
}
