% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{print.mccTable}
\alias{print.mccTable}
\title{Prints data into matched ca-co format}
\usage{
\method{print}{mccTable}(x, ...)
}
\arguments{
\item{x}{mccTable object}
}
\value{
NA
}
\description{
Prints new matrix represented as a table for a case control study
}
