% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{print.rateTable}
\alias{print.rateTable}
\title{Prints rate table}
\usage{
\method{print}{rateTable}(x, ...)
}
\arguments{
\item{x}{rate table object}
}
\value{
NA
}
\description{
Creates output of rate table
}
