% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{print.riskTable}
\alias{print.riskTable}
\title{Prints risk table}
\usage{
\method{print}{riskTable}(x, ...)
}
\arguments{
\item{x}{riskTable object}
}
\value{
NA
}
\description{
Prints new matrix represented as a table for a closed cohort (or cross-sectional)
}
