% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{summary.ccTable}
\alias{summary.ccTable}
\title{Summarizes data from ca-co format}
\usage{
\method{summary}{ccTable}(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{ccTable object}

\item{alpha}{alpha threshold}
}
\value{
NA
}
\description{
Creates new matrix represented as a table for a case control study
}
