% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{summary.mccTable}
\alias{summary.mccTable}
\title{Summarizes data into matched ca-co format}
\usage{
\method{summary}{mccTable}(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{mccTable object}

\item{alpha}{alpha threshold}
}
\value{
NA
}
\description{
Summarizes new matrix represented as a table for a case control study
}
