% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{summary.rateTable}
\alias{summary.rateTable}
\title{Prints summary rate table}
\usage{
\method{summary}{rateTable}(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{rate table object}

\item{alpha}{alpha threshold}
}
\value{
NA
}
\description{
Creates summary output of rate table
}
