% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_v2.R
\name{summary.riskTable}
\alias{summary.riskTable}
\title{Summarizes risk table}
\usage{
\method{summary}{riskTable}(x, alpha = 0.05, ...)
}
\arguments{
\item{x}{riskTable object}

\item{alpha}{alpha threshold}
}
\value{
NA
}
\description{
Summarizes new matrix represented as a table for a closed cohort (or cross-sectional)
}
